pyproject.toml
[project]
name = "agents"
version = "0.1.0"
description = "Enterprise-grade AI Agent platform with RAG, multiple LLM providers, and microservices architecture"
readme = "README.md"
requires-python = ">=3.10"
license = {text = "Apache-2.0"}
authors = [
    {name = "AI eUp", email = "your.email@example.com"}
]

dependencies = [
    # A2A SDK - Core
    "a2a-sdk>=0.3.11",
    
    # Web Framework
    "fastapi>=0.109.0",
    "uvicorn[standard]>=0.27.0",
    "starlette>=0.36.0",
    
    # LangChain Ecosystem
    "langchain>=0.1.0",
    "langchain-core>=0.1.0",
    "langchain-community>=0.0.20",
    "langgraph>=0.0.20",
    
    # LLM Providers
    "openai>=1.10.0",
    "anthropic>=0.18.0",
    "google-generativeai>=0.3.0",
    
    # Vector Databases
    "qdrant-client>=1.7.0",
    "chromadb>=0.4.22",
    
    # Embeddings
    "sentence-transformers>=2.3.0",
    
    # Data Processing
    "pydantic>=2.5.0",
    "pydantic-settings>=2.1.0",
    
    # Redis & Caching
    "redis>=5.0.1",
    "hiredis>=2.3.0",
    
    # Document Processing
    "pypdf>=4.0.0",
    "python-docx>=1.1.0",
    "python-multipart>=0.0.6",
    
    # Utilities
    "python-dotenv>=1.0.0",
    "httpx>=0.26.0",
    "aiofiles>=23.2.1",
    
    # Logging & Monitoring
    "structlog>=24.1.0",
    "python-json-logger>=2.0.7",
]

[project.optional-dependencies]
# A2A Extended Features
a2a-http = ["a2a-sdk[http-server]"]
a2a-grpc = ["a2a-sdk[grpc]"]
a2a-sql = ["a2a-sdk[sql]"]
a2a-all = ["a2a-sdk[all]"]

# Development
dev = [
    "pytest>=7.4.0",
    "pytest-asyncio>=0.23.0",
    "pytest-cov>=4.1.0",
    "black>=24.1.0",
    "ruff>=0.1.0",
    "mypy>=1.8.0",
    "pre-commit>=3.6.0",
]

# Production
prod = [
    "gunicorn>=21.2.0",
    "prometheus-client>=0.19.0",
    "opentelemetry-api>=1.22.0",
    "opentelemetry-sdk>=1.22.0",
    "opentelemetry-instrumentation-fastapi>=0.43b0",
]

[tool.poetry]
package-mode = false

[tool.black]
line-length = 100
target-version = ['py310', 'py311', 'py312']
include = '\.pyi?$'

[tool.ruff]
line-length = 100
target-version = "py310"
select = ["E", "F", "I", "N", "W", "UP"]
ignore = ["E501"]

[tool.mypy]
python_version = "3.10"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = false
ignore_missing_imports = true

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py", "*_test.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
asyncio_mode = "auto"

[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"